Implement RetroFX post-processing (Grayscale, Scanlines) (#215)
- Implemented `apply_grayscale` in `src/post_process.rs` using fixed-point arithmetic (`(77R+150G+29B) >> 8`) for efficient luminance calculation.
- Implemented `apply_scanlines` in `src/post_process.rs` using SWAR bitwise operations (`((p >> 1) & 0x7F7F7F7F)`) to darken odd rows without unpacking pixels.
- Added `tests/post_process_test.rs` to verify correctness.
- Added `benches/post_process.rs` to benchmark performance (~1.8ms for grayscale, ~290µs for scanlines at 1080p).
- Updated `.jules/nova.md` to mark RetroFX as Merged.
- Fixed Clippy warnings in `benches/simd_profiling.rs`.

+++ src/rasterizer.rs
@@ -1017,8 +1017,8 @@ fn draw_scanline_textured_perspective(
 
                 // Hoist texture properties
                 let tex_pixels = &texture.pixels;
-                let tex_w = texture.width as u32;
-                let tex_h = texture.height as u32;
+                let tex_w = texture.width;
+                let tex_h = texture.height;
                 let tex_w_usize = tex_w as usize;
 
                 let shift = texture.width_shift;
